Add patch to tolerate getrandom() returning EAGAIN.
If a process that uses Cryptography was started before the kernel
CRNG is fully initialised (e.g. during, or soon after, boot on a
system with no hardware RNG), then all subsequent attempts by the
process to read random bytes out of OpenSSL would fail. Change the
behaviour so that, instead of tainting the process forever,
subsequent attempts to read random bytes will check to see if the
CRNG has been initialised.